+++ apps/store/src/components/ProductPage/ProductPage.types.ts
@@ -1,7 +1,7 @@
 import type { ProductData } from '@/components/ProductData/ProductData.types'
+import type { AverageRating, ReviewComments } from '@/features/memberReviews/productReviews.types'
 import type { TrustpilotData } from '@/features/memberReviews/trustpilot.types'
 import { Template } from '@/services/PriceCalculator/PriceCalculator.types'
-import type { AverageRating, ReviewComments } from '@/services/productReviews/productReviews.types'
 import { ProductStory, StoryblokPageProps } from '@/services/storyblok/storyblok'
 
 export type ProductPageProps = StoryblokPageProps & {
